refactor(components/bnBtn): rename btn to bnBtn
